% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_attrs.R
\name{get_dta_cols}
\alias{get_dta_cols}
\title{Title}
\usage{
get_dta_cols(dat)
}
\arguments{
\item{dat}{a dataframe imported with col name attrs}
}
\value{
vector of column names and attrs
}
\description{
Title
}
