% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prop_CI_cols.R
\name{prop_ci_cols}
\alias{prop_ci_cols}
\title{add_prop_CI)cold}
\usage{
prop_ci_cols(
  .data,
  numerator = "n",
  denominator = "d",
  conf.level = 0.95,
  method = "wald",
  rand = 123,
  R = 9999,
  bootci.type = "all",
  alternative = c("two.sided", "less", "greater"),
  ...
)
}
\arguments{
\item{.data}{dataframe to add columns to}

\item{numerator}{the name of the numerator column}

\item{denominator}{the name of the denominator column}

\item{conf.level}{confidence level for CIs}

\item{method}{type of CI to be calculated}

\item{rand}{random seed}

\item{R}{output format}
}
\value{
three dataframe columns
}
\description{
add_prop_CI)cold
}
